test(runner): move the disclosure tests to the unix-only file
Both need a sealed execution object, which Windows declines by design
(FILE_EXEC_UNSUPPORTED before the file is opened), so they belonged beside
the other tests that exercise the lane rather than in the cross-platform
file.
