+++ pkg/runner/file_exec_unix_test.go
@@ -3,9 +3,12 @@
 package runner
 
 import (
+	"crypto/sha256"
+	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strings"
 	"testing"
 
 	"github.com/alpacax/alpamon/v2/pkg/executor/handlers/common"
@@ -106,4 +109,42 @@ func TestOpenVerifiedFile_SurvivesDeletionOfTheOriginal(t *testing.T) {
 	require.NoError(t, os.Remove(path))
 
 	assert.Equal(t, "ORIGINAL\n", runVerifiedFile(t, sealed))
+}
+
+// A refusal is reported to the requester as the command result, and the
+// requester chose the path. Returning what the file hashed to would make the
+// refusal itself a disclosure: the assessor judges the submitted content, so a
+// benign script paired with a path like /etc/shadow auto-approves, mismatches,
+// and hands back the digest of a file the requester may not read.
+func TestOpenVerifiedFile_MismatchNeverReportsTheObservedDigest(t *testing.T) {
+	secret := []byte("root-only secret\n")
+	path := filepath.Join(t.TempDir(), "secret")
+	require.NoError(t, os.WriteFile(path, secret, 0o600))
+	observed := fmt.Sprintf("%x", sha256.Sum256(secret))
+	approved := strings.Repeat("b", 64)
+
+	_, err := openVerifiedFile(path, approved)
+
+	require.Error(t, err)
+	require.ErrorIs(t, err, errHashMismatch)
+	assert.NotContains(t, err.Error(), observed,
+		"the refusal must not disclose what the file hashed to")
+	// The expected digest came from the requester, so echoing it tells them
+	// nothing they did not already send.
+	assert.Contains(t, err.Error(), approved)
+}
+
+// The operator still needs the real value to diagnose a genuine mismatch; it
+// travels on the error type for the local log rather than in Error().
+func TestHashMismatchError_CarriesTheObservedDigestForTheLog(t *testing.T) {
+	secret := []byte("root-only secret\n")
+	path := filepath.Join(t.TempDir(), "secret")
+	require.NoError(t, os.WriteFile(path, secret, 0o600))
+	observed := fmt.Sprintf("%x", sha256.Sum256(secret))
+
+	_, err := openVerifiedFile(path, strings.Repeat("b", 64))
+
+	var mismatch *hashMismatchError
+	require.ErrorAs(t, err, &mismatch)
+	assert.Equal(t, observed, mismatch.Observed())
 }
